fix: Downgrade single digit division from uint64_t to uint32_t
When dividing the uint_64 into a lower and upper part, there was an implicit assumption that the result would fit in 32 bits as the lower and the upper parts were shifted by 32 bits.

If the reminder was larger than 32 bits, the shift would cut off all bits larger than 32 causing the result to be incorrect.